% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpops.R
\name{df2sf}
\alias{df2sf}
\title{Creates a simple features (sf) object from a dataframe with latitude and longitude values}
\usage{
df2sf(dataframe, longitude = "decimalLongitude",
  latitude = "decimalLatitude", default_crs = "+proj=longlat +datum=WGS84")
}
\arguments{
\item{dataframe}{dataframe with latitude and longitude}

\item{longitude}{fieldname that contains the longitude values}

\item{latitude}{fieldname that contains the latitude values}

\item{default_crs}{coordinate reference system}
}
\value{
simple feature object
}
\description{
Creates a simple features (sf) object from a dataframe with latitude and longitude values
}
