% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpops.R
\name{flattenSquid}
\alias{flattenSquid}
\title{Flattens the squidpop data to a singlefile for archiving and analysis}
\usage{
flattenSquid(df_tabs)
}
\arguments{
\item{df_tabs}{list of dataframes}
}
\value{
r dataframe
}
\description{
Joins the data and locations tabs using the locationIDs and saves result
}
