% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpops.R
\name{packageSquid}
\alias{packageSquid}
\title{Writes the squidpop xlsx to an eml file}
\usage{
packageSquid(xlsx, csv_file, eml_file)
}
\arguments{
\item{xlsx}{path to the input squidpop excel workbook}

\item{csv_file}{path to the export flattend csv}

\item{eml_file}{path to the export eml}
}
\value{
simple feature object
}
\description{
Writes the squidpop xlsx to an eml file
}
