% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squidpops.R
\name{readSquid}
\alias{readSquid}
\title{Loads the squidpop template into R as a list of dataframes}
\usage{
readSquid(xlsx)
}
\arguments{
\item{xlsx}{path to the squidpop excel workbook}
}
\value{
r dataframe
}
\description{
Imports all the tabs in the squidpop template into a list of r dataframes
}
